backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
supabase
fastapi-csrf-protect